docs: transfer the rouletteSeries constant to a separate module
