+++ roulette/models/trackModel.js
@@ -1,7 +1,7 @@
 // import "./menuModel.js";
 // import { minBet } from "./menuModel.js";
 // import { maxBet } from "./menuModel.js";
-import { rouletteSeries } from "./menuModel.js";
+import { rouletteSeries } from "./consts/rouletteSeries.js";
 import { rltPos } from "./consts/rltPos.js";
 import { payoutRatios } from "./consts/payoutRatios.js";
 // import { seriesCalc } from "./menuModel.js";
